use correct returntype on CloudFront::purge
FOSHttpCache 3 adds return types. 

if we would use the `self` return type on the interface, this would only declare that it returns an instance of the interface. `static` means its the exact class.

php explodes when it sees an extending class use `self`, so we should use `static` here. (although this being a final class, it happens to have the exact same semantics i think)